refactor/#86: project, projectUser 폴더 분리
